Updated logging to give it functionality
Added small logging functionality. This will create a very long string of data when the methods are used. Working on making the data tree more readable. (Visual studio isn't showing me errors in the code so sorry for any errors). There may be some issues with some data trying to be written at the same time